ml4pl: Split multiprocessing run_id tests into separate test.
