+++ deeplearning/ml4pl/run_id_test.py
@@ -15,9 +15,6 @@
 # limitations under the License.
 """Unit tests for //deeplearning/ml4pl:run_id."""
 import datetime
-import multiprocessing
-import subprocess
-import sys
 import time
 
 import sqlalchemy as sql
@@ -126,43 +123,6 @@ def test_GenerateGlobalUnique_unique_timestamps():
     # Check that run ID has changed.
     assert run_id_ != previous_run_id
     previous_run_id = run_id_
-
-
-def _MakeARunId(*args):
-  """Generate a run ID."""
-  return run_id.RunId.GenerateGlobalUnique()
-
-
-def test_GenerateGlobalUnique_subprocess():
-  """Run IDs should be unique when running as subprocesses."""
-  unique_run_ids = set()
-
-  for _ in range(30):
-    run_id_ = subprocess.check_output(
-      [
-        sys.executable,
-        "-c",
-        "from deeplearning.ml4pl import run_id; "
-        "print(run_id.RunId.GenerateGlobalUnique())",
-      ],
-      universal_newlines=True,
-    )
-
-    if run_id_ in unique_run_ids:
-      test.Fail("Non-unique run ID generated")
-    unique_run_ids.add(run_id_)
-
-
-def test_GenerateGlobalUnique_multiprocessed():
-  """Generate a bunch of run IDs concurrently and check that they are unique."""
-  unique_run_ids = set()
-
-  with multiprocessing.Pool() as p:
-    for run_id_ in p.imap_unordered(_MakeARunId, range(30)):
-      print("Generated run ID:", run_id)
-      if run_id_ in unique_run_ids:
-        test.Fail("Non-unique run ID generated")
-      unique_run_ids.add(run_id_)
 
 
 if __name__ == "__main__":
